examples/apps/pipecat-chatbot/pyproject.toml
[tool.poetry]
name = "pipecat-chatbot"
version = "0.1.0"
description = ""
authors = ["Andrew Chang <andrew@acalabs.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pipecat-ai = {extras = ["daily", "openai", "silero"], version = "^0.0.36"}
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
twilio = "^9.2.3"
pinecone = "^4.0.0"
elixir-ai = { path = "../../..", develop = true }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
